Refactor move refreshTooltips to project js, remove copyText
