+++ commcare_connect/static/js/project.js
@@ -1,3 +1,14 @@
 import '../sass/project.scss';
+import * as bootstrap from 'bootstrap';
 
 /* Project specific Javascript goes here. */
+
+function refreshTooltips() {
+  const tooltipTriggerList = document.querySelectorAll(
+    '[data-bs-toggle="tooltip"]',
+  );
+  const tooltipList = [...tooltipTriggerList].map(
+    (tooltipTriggerEl) => new bootstrap.Tooltip(tooltipTriggerEl),
+  );
+}
+window.refreshTooltips = refreshTooltips;
